+++ packages/gamut-labs/src/GlobalFooter/__tests__/GlobalFooter-test.tsx
@@ -4,7 +4,6 @@ import { GlobalFooter } from '..';
 
 const renderView = setupRtl(GlobalFooter, {
   onClick: jest.fn(),
-  showNewFooter: false,
   userGeo: 'US',
 });
 
@@ -13,19 +12,5 @@ describe('GlobalFooter', () => {
     const { view } = renderView();
 
     view.getByRole('contentinfo');
-  });
-
-  it('shows new footer when showNewFooter is true', () => {
-    const { view } = renderView({ showNewFooter: true });
-
-    view.getByTitle('Codecademy from Skillsoft');
-    expect(view.queryByText('Company')).toBeNull();
-  });
-
-  it('does not show new footer when showNewFooter is false', () => {
-    const { view } = renderView();
-
-    view.getByText('Company');
-    expect(view.queryByTitle('Codecademy from Skillsoft')).toBeNull();
   });
 });