+++ packages/gamut-labs/src/GlobalFooter/index.tsx
@@ -26,21 +26,13 @@ export type GlobalFooterProps = {
   onMadeInClick?: (text: string) => void;
 
   /**
-   * If new footer should be shown
-   */
-  showNewFooter: boolean;
-
-  /**
    * Geographic region of the user viewing the footer, such as "IN" or "US".
    */
   userGeo: string;
 };
 
-// To be deleted after footer launches - using React.Context here will be the easiest cleanup.
-export const FooterFFContext = React.createContext({ showNewFooter: false });
-
 const FooterContainer = styled.footer`
-  border-top: 1px solid ${themed('colors.navy')};
+  border-top: 1px solid ${themed('colors.secondary')};
 
   @media print {
     display: none;
@@ -52,21 +44,18 @@ export const GlobalFooter: React.FC<GlobalFooterProps> = ({
   hidePricing,
   onClick,
   onMadeInClick,
-  showNewFooter,
   userGeo,
 }) => {
   return (
-    <FooterFFContext.Provider value={{ showNewFooter }}>
-      <FooterContainer className={className} role="contentinfo">
-        <ContentContainer>
-          <FooterNavLinks
-            hidePricing={hidePricing}
-            onClick={onClick}
-            userGeo={userGeo}
-          />
-        </ContentContainer>
-        <FooterLegal onClick={onClick} onMadeInClick={onMadeInClick} />
-      </FooterContainer>
-    </FooterFFContext.Provider>
+    <FooterContainer className={className} role="contentinfo">
+      <ContentContainer>
+        <FooterNavLinks
+          hidePricing={hidePricing}
+          onClick={onClick}
+          userGeo={userGeo}
+        />
+      </ContentContainer>
+      <FooterLegal onClick={onClick} onMadeInClick={onMadeInClick} />
+    </FooterContainer>
   );
 };